Move ru.d3rvich.browse into ru.d3rvich.feature.browse.
