+++ app/src/main/java/ru/d3rvich/jetgames/navigation/MainContent.kt
@@ -26,7 +26,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import com.arkivanov.decompose.extensions.compose.stack.Children
 import com.arkivanov.decompose.extensions.compose.subscribeAsState
-import ru.d3rvich.browse.BrowseScreen
+import ru.d3rvich.feature.browse.BrowseScreen
 import ru.d3rvich.favorites.FavoritesScreen
 import ru.d3rvich.home.HomeScreen
 import ru.d3rvich.jetgames.navigation.main.MainComponent
