+++ feature/browse/src/main/java/ru/d3rvich/feature/browse/BrowseScreen.kt
@@ -1,4 +1,4 @@
-package ru.d3rvich.browse
+package ru.d3rvich.feature.browse
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
@@ -14,9 +14,10 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import org.koin.compose.viewmodel.koinViewModel
-import ru.d3rvich.browse.model.BrowseUiState
-import ru.d3rvich.browse.views.GenresView
-import ru.d3rvich.browse.views.PlatformsView
+import ru.d3rvich.browse.R
+import ru.d3rvich.feature.browse.model.BrowseUiState
+import ru.d3rvich.feature.browse.views.GenresView
+import ru.d3rvich.feature.browse.views.PlatformsView
 
 /**
  * Created by Ilya Deryabin at 05.06.2024
