+++ feature/browse/src/main/java/ru/d3rvich/feature/browse/views/GenresView.kt
@@ -1,4 +1,4 @@
-package ru.d3rvich.browse.views
+package ru.d3rvich.feature.browse.views
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.PaddingValues
